bla bla

+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: ppourraj <ppourraj@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/15 19:35:59 by ppourraj          #+#    #+#             */
-/*   Updated: 2026/04/20 17:06:00 by ppourraj         ###   ########.fr       */
+/*   Updated: 2026/04/24 16:48:25 by ppourraj         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	if (n == 0)
 		return (0);
